move shared code out of receiverwrapper and into domain/decorators.py
