+++ corehq/apps/domain/decorators.py
@@ -2,6 +2,7 @@
 import base64
 from functools import wraps
 import logging
+import re
 
 # Django imports
 from django.conf import settings
@@ -26,7 +27,6 @@ from tastypie.http import HttpUnauthorized
 # CCHQ imports
 from corehq.apps.domain.models import Domain
 from corehq.apps.domain.utils import normalize_domain_name
-from corehq.apps.receiverwrapper.util import determine_authtype_from_user_agent
 from corehq.apps.users.models import CouchUser
 from corehq import privileges
 
@@ -177,6 +177,27 @@ def login_or_basic_ex(allow_cc_users=False):
     return _login_or_challenge(basicauth(), allow_cc_users=allow_cc_users)
 
 login_or_basic = login_or_basic_ex()
+
+
+J2ME = 'j2me'
+ANDROID = 'android'
+
+
+def guess_phone_type_from_user_agent(user_agent):
+    """
+    A really dumb utility that guesses the phone type based on the user-agent header.
+    """
+    j2me_pattern = '[Nn]okia|NOKIA|CLDC|cldc|MIDP|midp|Series60|Series40|[Ss]ymbian|SymbOS|[Mm]aemo'
+    return J2ME if user_agent and re.search(j2me_pattern, user_agent) else ANDROID
+
+
+def determine_authtype_from_user_agent(request):
+    user_agent = request.META.get('HTTP_USER_AGENT')
+    type_to_auth_map = {
+        J2ME: 'digest',
+        ANDROID: 'basic',
+    }
+    return type_to_auth_map[guess_phone_type_from_user_agent(user_agent)]
 
 
 def login_or_digest_or_basic(fn):
