+++ corehq/apps/domain/tests/__init__.py
@@ -5,3 +5,4 @@ from .test_utils import *
 from .test_domain_name_generation import *
 from .test_domain_transfer import *
 from .test_delete_domain import *
+from .test_guess_phone_type import *
